// lib/portfolio.generated.ts
export interface PortfolioSpec { [key: string]: string }

export interface PortfolioWork {
  id: number;
  slug: string;
  title: string;
  category: string;
  categorySlug: string;
  coverImage: string;
  images: string[];
  description?: string;
  specs?: PortfolioSpec;
}